backend/requirements.txt
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloths-segmentation==0.0.2
colorama==0.4.6
distro==1.9.0
docker-pycreds==0.4.0
efficientnet_pytorch==0.7.1
eval_type_backport==0.2.0
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.36.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jpeg4py==0.1.4
lazy_loader==0.4
lightning-utilities==0.11.8
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
networkx==3.4.2
numpy==2.1.3
openai==1.54.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-toolbelt==0.6.3
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segmentation-models-pytorch==0.3.4
sentry-sdk==2.18.0
setproctitle==1.3.4
setuptools==75.5.0
simsimd==6.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
stringzilla==3.10.10
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.7
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wandb==0.18.7
yarl==1.17.2
